fix: split execution claim lifecycle tests
